refactor: Consolidate web-ui test files to __tests__/ directory
Migrate all test files from src/ to __tests__/ for single source of truth:

- Merge AgentCard tests (28 comprehensive + 3 navigation tests)
- Move component tests: DiscoveryProgress, PhaseIndicator, ProgressBar,
  ProjectCreationForm, Spinner, PRDModal, TaskTreeView
- Move lib tests: api.test.ts
- Remove duplicate HomePage/ProjectList tests (kept in __tests__/)
- Update imports to use @/ alias pattern
- Delete all test files from src/

Result: 50 test suites, 1300 tests (was 1546 with duplicates)
